Updated the workflow for option 1 in Menu class, updated ProductApplicationService class and Product class.

+++ app/src/main/java/com/mightyfour/domain/Product.java
@@ -28,7 +28,7 @@ public class Product {
     private Category category;
     private LIFESPAN lifespan;
 
-    public Product(String product_name, Material material, int duration){
+    public Product(String product_name, Material material, double duration){
 
         this.product_name = product_name;
